MyWork component created + mobile view
