+++ src/reusable/MyWork/MyWork.css
@@ -1,0 +1,35 @@
+.my_work__container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.my_work__title {
+  margin-top: 48px;
+}
+
+.case_study__container {
+  margin-top: 24px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 64px;
+}
+
+@media (min-width: 1008px) {
+  .my_work__container {
+    max-width: 1324px;
+    padding: 16px;
+  }
+
+  .my_work__title {
+    margin-top: 64px;
+  }
+
+  .case_study__container {
+    margin-top: 24px;
+    flex-direction: row;
+    flex-wrap: wrap;
+    gap: 84px 224px;
+  }
+}
