+++ src/reusable/MyWork/types.ts
@@ -1,0 +1,5 @@
+import { caseStudiesData } from "../../data/CaseStudies";
+
+export interface MyWorkProps {
+  caseStudies: typeof caseStudiesData.caseStudies;
+}
